moved logic to behavioral trait
