+++ src/Doctrine/Behavior/ModifiedAt.php
@@ -7,7 +7,7 @@
  * file that was distributed with this source code.
  */
 
-namespace App\Doctrine;
+namespace App\Doctrine\Behavior;
 
 interface ModifiedAt
 {
