+++ src/Entity/Timesheet.php
@@ -9,7 +9,8 @@
 
 namespace App\Entity;
 
-use App\Doctrine\ModifiedAt;
+use App\Doctrine\Behavior\ModifiedAt;
+use App\Doctrine\Behavior\ModifiedTrait;
 use App\Validator\Constraints as Constraints;
 use DateTime;
 use DateTimeZone;
@@ -48,6 +49,8 @@ use Symfony\Component\Validator\Constraints as Assert;
 #[Constraints\TimesheetDeactivated]
 class Timesheet implements EntityWithMetaFields, ExportableItem, ModifiedAt
 {
+    use ModifiedTrait;
+
     /**
      * Category: Normal work-time (default category)
      */
@@ -190,8 +193,6 @@ class Timesheet implements EntityWithMetaFields, ExportableItem, ModifiedAt
     #[ORM\Column(name: 'category', type: 'string', length: 10, nullable: false, options: ['default' => 'work'])]
     #[Assert\NotNull]
     private ?string $category = self::WORK;
-    #[ORM\Column(name: 'modified_at', type: 'datetime_immutable', nullable: true)]
-    private \DateTimeImmutable $modifiedAt;
     /**
      * Tags
      *
@@ -220,7 +221,7 @@ class Timesheet implements EntityWithMetaFields, ExportableItem, ModifiedAt
     {
         $this->tags = new ArrayCollection();
         $this->meta = new ArrayCollection();
-        $this->modifiedAt = new \DateTimeImmutable('now', new \DateTimeZone('UTC'));
+        $this->setModifiedAt(new \DateTimeImmutable('now', new \DateTimeZone('UTC')));
     }
 
     /**
@@ -550,16 +551,6 @@ class Timesheet implements EntityWithMetaFields, ExportableItem, ModifiedAt
         return $this;
     }
 
-    public function getModifiedAt(): \DateTimeImmutable
-    {
-        return $this->modifiedAt;
-    }
-
-    public function setModifiedAt(\DateTimeImmutable $dateTime): void
-    {
-        $this->modifiedAt = $dateTime;
-    }
-
     /**
      * @return Collection|MetaTableTypeInterface[]
      */
@@ -607,7 +598,7 @@ class Timesheet implements EntityWithMetaFields, ExportableItem, ModifiedAt
     {
         // this needs to be done, otherwise doctrine will not see the item as changed
         // and the calculators will not run
-        $this->modifiedAt = new \DateTimeImmutable('now', new \DateTimeZone('UTC'));
+        $this->setModifiedAt(new \DateTimeImmutable('now', new \DateTimeZone('UTC')));
 
         if (null === ($current = $this->getMetaField($meta->getName()))) {
             $meta->setEntity($this);
